pyproject.toml
[project]
name = "pharma-derive"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.10,<3",
    "pydantic-ai>=1.0,<2",
    "pandas>=2.2,<3",
    "networkx>=3.4,<4",
    "pyyaml>=6.0,<7",
    "loguru>=0.7,<1",
    "python-statemachine>=3.0,<4",
    "sqlalchemy[asyncio]>=2.0,<3",
    "aiosqlite>=0.21,<1",
    "pyreadstat>=1.3,<2",
    "streamlit>=1.40,<2",
    "pydantic-settings>=2.7,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.24,<1",
    "pytest-cov>=6.0,<7",
    "ruff>=0.8,<1",
    "pyright>=1.1,<2",
    "pandas-stubs>=2.2,<3",
    "pre-commit>=4.5,<5",
    "import-linter>=2.11,<3",
    "vulture>=2.16,<3",
    "radon>=6.0,<7",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH", "RUF", "S", "T20", "C4", "RET"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]  # assert is expected in tests
"src/domain/dag.py" = ["S101"]  # assert guards cached computation invariant

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"
reportMissingTypeStubs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
omit = ["src/ui/*"]  # Streamlit pages are integration-tested manually, not unit-testable

[dependency-groups]
dev = [
    "pydot>=4.0.1",
]
